+++ src/game.h
@@ -7,11 +7,11 @@ extern "C" {
 
 #include <stdint.h>
 
-uint8_t draw(uint8_t, int16_t, int16_t, uint8_t, int8_t, int16_t, int16_t);
+uint8_t draw(uint8_t, int,  int,  uint8_t, int8_t, int,  int);
 
-int16_t eat(int16_t, int16_t, int16_t);
+int eat(int,  int,  int);
 
-void youLose(int16_t);
+void youLose(int);
 
 #ifdef __cplusplus
 }
